backend/requirements.txt
fastapi
groq
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
email-validator
requests
wikipedia
langchain-community
pytest
httpx
python-multipart
google-genai
